pdf-ingestion-service/requirements.txt
# FastAPI and server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# HTTP client for future use
httpx==0.26.0

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# S3-compatible client for Cloudflare R2
boto3==1.34.34

# PDF parsing
pypdf==4.0.1

# Vector database
qdrant-client==1.7.3

# Local embeddings
sentence-transformers==2.3.1
